Add separate word for special characters
Special characters used to be part of words, eg 'street,' so comma is considered part of the word and so needs to be typed as well as the letters to finish the word. Now this is not the case, although 'street,' is still considered the same word when replaced, when 'street' is typed, 'street' completes, leaving ',' as '_', until that is typed as well.