Add poem name to poem text files
Poem text files should have the entire poem content in them, including title and author. Title of poem now goes in name of poem text file and on the first line of the file, with a blank line before the rest of the poem content. If the title in the filename doesn't match this file, an error is thrown in modifyRawPoemsJSON.js, as a sanity check.