Add associated text modification for notes
Now all words have an event handler that adds it to highlighted text and highlights it, or the opposite if already highlighted. When a POST request is sent for a note to be updated, highlightedText is used in the newVersion object, which updates the text that is associated with a note text in the convertedPoems.json file.