Bug fix
When total number of words in poem being changed to was less than the number of words that were trying to be removed, error occurred. Fix by ensuring the number of words to remove is not larger than the number of words in the poem, when the poem is changed. Became broken in last commit when the rangeValidationForNumberOfWordsToReplace function was removed from the replaceWords function.