Center text in input when input value length > 0
Letters such as 'i' don't look good when the input's 'textAlign' is set to 'start'. However, if the input's value is centered in CSS, the cursor is in the middle of the input box, which doesn't look good. Solution is to change the textAlign property in javascript depending on whether the input has a letter in it or not.