Add conflict check
If the content of a poem is changed, quotes and notes containing a word that is edited or removed, will no longer be valid. To solve this, when convertedPoems is being updated, it checks all the notes and quotes and removes all notes and quotes that are no longer valid. All invalid notes and quotes are moved to a separate file called invalidNotesAndQuotes.json, where the admin can then look through and try to restore these broken quotes. In the future, these will appear in the preprocessing UI as a checklist making them possible to restore without looking in the files manually.